requirements.txt
langchain_google_genai
langchain_community
python-dotenv
langchain-core
langchain-ollama
pydantic
# LangChain Core
langchain
langchain-core
grandalf
# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn